ai/ai-vision/ai-vision-and-sam/files/requirements.txt
huggingface-hub==0.35.0
numpy==2.2.6
oci==2.160.1
opencv-python==4.12.0.88
pandas==2.3.2
pillow==11.3.0
python-dotenv==1.1.1
sam2==1.1.0
streamlit==1.49.1
torch==2.8.0
torchvision==0.23.0